Move caches from internal -> cache package
No reason for them to be internal in hindsight

Also ensure a java 8 compatible build, the if statement doesn't seem to
work as you'd expect
