+++ build.gradle
@@ -17,13 +17,9 @@ repositories {
 }
 
 compileJava {
-	if (JavaVersion.current() <= JavaVersion.VERSION_1_8) {
-		// TODO BUILD remove when we no longer support java 8
-		java.sourceCompatibility = JavaVersion.VERSION_1_8
-		java.targetCompatibility = JavaVersion.VERSION_1_8
-	} else {
-		options.release = 8
-	}
+	// TODO BUILD remove when we no longer support java 8, use `options.release = 11 if necessary`
+	java.sourceCompatibility = JavaVersion.VERSION_1_8
+	java.targetCompatibility = JavaVersion.VERSION_1_8
 }
 
 test {
