+++ src/test/java/us/kbase/test/auth/client/cache/TokenCacheTest.java
@@ -1,4 +1,4 @@
-package us.kbase.test.auth.client.internal;
+package us.kbase.test.auth.client.cache;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertNotNull;
@@ -14,7 +14,7 @@ import java.util.stream.IntStream;
 import org.junit.Test;
 
 import us.kbase.auth.AuthToken;
-import us.kbase.auth.client.internal.TokenCache;
+import us.kbase.auth.client.cache.TokenCache;
 import us.kbase.test.common.TestCommon;
 
 public class TokenCacheTest {
